fix(adblocker): split logic into separate files
