+++ src/background/adblocker/index.js
@@ -9,616 +9,5 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0
  */
 
-import { store } from 'hybrids';
-import { filterRequestHTML, updateResponseHeadersWithCSP } from '@ghostery/adblocker-webextension';
-import scriptlets from '@ghostery/scriptlets';
-import { FLAG_SUBFRAME_SCRIPTING } from '@ghostery/config';
-
-import { resolveFlag } from '/store/config.js';
-import Options, { ENGINES, getPausedDetails } from '/store/options.js';
-
-import { isWebkit } from '/utils/browser-info.js';
-import * as exceptions from '/utils/exceptions.js';
-import * as engines from '/utils/engines.js';
-import * as trackerdb from '/utils/trackerdb.js';
-import * as OptionsObserver from '/utils/options-observer.js';
-import Request, { parseWithCache } from '/utils/request.js';
-import asyncSetup from '/utils/setup.js';
-
-import { updateDNRRulesForExceptions } from '../exceptions.js';
-import { tabStats, updateTabStats } from '../stats.js';
-import { getRedirectProtectionUrl } from '../redirect-protection.js';
-
-import { FramesHierarchy } from './ancestors.js';
-
-const contentScripts = (() => {
-  const map = new Map();
-  return {
-    async register(hostname, code) {
-      this.unregister(hostname);
-      try {
-        const contentScript = await browser.contentScripts.register({
-          js: [
-            {
-              code,
-            },
-          ],
-          allFrames: true,
-          matches: [`https://*.${hostname}/*`, `http://*.${hostname}/*`],
-          matchAboutBlank: true,
-          matchOriginAsFallback: true,
-          runAt: 'document_start',
-          world: 'MAIN',
-        });
-        map.set(hostname, contentScript);
-      } catch (e) {
-        console.warn(e);
-        this.unregister(hostname);
-      }
-    },
-    isRegistered(hostname) {
-      return map.has(hostname);
-    },
-    unregister(hostname) {
-      const contentScript = map.get(hostname);
-      if (contentScript) {
-        contentScript.unregister();
-        map.delete(hostname);
-      }
-    },
-    unregisterAll() {
-      for (const hostname of map.keys()) {
-        this.unregister(hostname);
-      }
-    },
-  };
-})();
-
-function getEnabledEngines(config) {
-  if (config.terms) {
-    const list = ENGINES.filter(({ key }) => config[key]).map(({ name }) => name);
-
-    if (config.regionalFilters.enabled) {
-      list.push(...config.regionalFilters.regions.map((id) => `lang-${id}`));
-    }
-
-    if (config.fixesFilters && list.length) {
-      list.push(engines.FIXES_ENGINE);
-    }
-
-    list.push(engines.ELEMENT_PICKER_ENGINE);
-
-    if (config.customFilters.enabled) {
-      list.push(engines.CUSTOM_ENGINE);
-    }
-
-    return list;
-  }
-
-  return [];
-}
-
-function pause(ms) {
-  return new Promise((resolve) => setTimeout(resolve, ms));
-}
-
-export async function reloadMainEngine() {
-  // Delay the reload to avoid UI freezes in Firefox and Safari
-  if (__FIREFOX__ || isWebkit()) await pause(1000);
-
-  const options = await store.resolve(Options);
-  const enabledEngines = getEnabledEngines(options);
-
-  const resolvedEngines = (
-    await Promise.all(
-      enabledEngines.map((id) =>
-        engines
-          .init(id)
-          .catch(() => {
-            console.error(`[adblocker] failed to load engine: ${id}`);
-            return null;
-          })
-          .then((engine) => {
-            if (!engine) {
-              enabledEngines.splice(enabledEngines.indexOf(id), 1);
-            }
-            return engine;
-          }),
-      ),
-    )
-  ).filter((engine) => engine);
-
-  if (resolvedEngines.length) {
-    engines.replace(engines.MAIN_ENGINE, resolvedEngines);
-
-    console.info(`[adblocker] Main engine reloaded with: ${enabledEngines.join(', ')}`);
-  } else {
-    await engines.create(engines.MAIN_ENGINE);
-    console.info('[adblocker] Main engine reloaded with no filters');
-  }
-
-  if (__FIREFOX__) {
-    contentScripts.unregisterAll();
-  }
-}
-
-let updating = false;
-export async function updateEngines({ cache = true } = {}) {
-  if (updating) return;
-
-  try {
-    updating = true;
-
-    const options = await store.resolve(Options);
-    const enabledEngines = getEnabledEngines(options);
-
-    if (enabledEngines.length) {
-      let updated = false;
-
-      // Update engines from the list of enabled engines
-      await Promise.all(
-        enabledEngines.filter(engines.isPersistentEngine).map(async (id) => {
-          await engines.init(id);
-          updated = (await engines.update(id, { cache })) || updated;
-        }),
-      );
-
-      // Update TrackerDB engine
-      trackerdb.setup.pending && (await trackerdb.setup.pending);
-      const trackerdbUpdated = await engines.update(engines.TRACKERDB_ENGINE, { cache });
-
-      if (__CHROMIUM__ && trackerdbUpdated) {
-        // We need to reload DNR rules for exceptions if TrackerDB engine is updated,
-        // as rules rely on TrackerDB metadata
-        await updateDNRRulesForExceptions();
-      }
-
-      // Update timestamp after the engines are updated
-      await store.set(Options, { filtersUpdatedAt: Date.now() });
-
-      if (updated) await reloadMainEngine();
-    }
-  } finally {
-    updating = false;
-  }
-}
-
-export const UPDATE_ENGINES_DELAY = 60 * 60 * 1000; // 1 hour
-export const setup = asyncSetup('adblocker', [
-  OptionsObserver.addListener(async function adblockerEngines(options, lastOptions) {
-    const enabledEngines = getEnabledEngines(options);
-    const lastEnabledEngines = lastOptions && getEnabledEngines(lastOptions);
-
-    // Enabled engines changed (they might contain outdated filters)
-    const enginesChanged =
-      lastEnabledEngines &&
-      (enabledEngines.length !== lastEnabledEngines.length ||
-        enabledEngines.some((id, i) => id !== lastEnabledEngines[i]));
-
-    // Reload main engine:
-    // * when engine is not initialized or initialize fails (adblocker version mismatch)
-    // * when enabled engines changed
-    if (!(await engines.init(engines.MAIN_ENGINE)) || enginesChanged) {
-      await reloadMainEngine();
-    }
-
-    // Update engine filters:
-    // * when engines changed, so there might be re-enabled engines with outdated filters
-    // * when filters are outdated (older than 1 hour)
-    if (enginesChanged || options.filtersUpdatedAt < Date.now() - UPDATE_ENGINES_DELAY) {
-      await updateEngines();
-    }
-  }),
-]);
-
-function resolveInjectionTarget(details) {
-  const target = { tabId: details.tabId };
-
-  if (__CHROMIUM__ && details.documentId) {
-    target.documentIds = [details.documentId];
-  } else {
-    target.frameIds = [details.frameId];
-  }
-
-  return target;
-}
-
-const scriptletGlobals = {
-  // Request a real extension resource to obtain a dynamic ID to the resource.
-  // Redirect resources are defined with `use_dynamic_url` restriction.
-  // The dynamic ID is generated per session.
-  // refs https://developer.chrome.com/docs/extensions/reference/manifest/web-accessible-resources#manifest_declaration
-  warOrigin: chrome.runtime.getURL('/rule_resources/redirects/empty').slice(0, -6),
-};
-
-function injectScriptlets(filters, hostname, details) {
-  let contentScript = '';
-  for (const filter of filters) {
-    const parsed = filter.parseScript();
-
-    if (!parsed) {
-      console.warn('[adblocker] could not inject script filter:', filter.toString());
-      continue;
-    }
-
-    const scriptletName = `${parsed.name}${parsed.name.endsWith('.js') ? '' : '.js'}`;
-    const scriptlet = scriptlets[scriptletName];
-
-    if (!scriptlet) {
-      console.warn('[adblocker] unknown scriptlet with name:', scriptletName);
-      continue;
-    }
-
-    const func = scriptlet.func;
-    const args = [scriptletGlobals, ...parsed.args.map((arg) => decodeURIComponent(arg))];
-
-    if (__FIREFOX__) {
-      if (filter.hasSubframeConstraint()) {
-        contentScript += `window.parent!==window&&`;
-      }
-      contentScript += `(${func.toString()})(...${JSON.stringify(args)});\n`;
-      continue;
-    }
-
-    chrome.scripting.executeScript(
-      {
-        injectImmediately: true,
-        world: chrome.scripting.ExecutionWorld?.MAIN ?? (__FIREFOX__ ? undefined : 'MAIN'),
-        target: resolveInjectionTarget(details),
-        func,
-        args,
-      },
-      () => {
-        if (chrome.runtime.lastError) {
-          console.warn(chrome.runtime.lastError);
-        }
-      },
-    );
-  }
-
-  if (__FIREFOX__) {
-    if (filters.length === 0) {
-      contentScripts.unregister(hostname);
-    } else if (!contentScripts.isRegistered(hostname)) {
-      contentScripts.register(hostname, contentScript);
-    } else {
-      // do nothing if already registered
-    }
-  }
-}
-
-function injectStyles(styles, details) {
-  chrome.scripting
-    .insertCSS({
-      css: styles,
-      origin: 'USER',
-      target: resolveInjectionTarget(details),
-    })
-    .catch((e) => console.warn('[adblocker] failed to inject CSS', e));
-}
-
-const SUBFRAME_SCRIPTING = resolveFlag(FLAG_SUBFRAME_SCRIPTING);
-
-let framesHierarchy;
-if (__CHROMIUM__) {
-  framesHierarchy = new FramesHierarchy();
-
-  framesHierarchy.handleWebWorkerStart();
-  framesHierarchy.handleWebextensionEvents();
-}
-/*
- * returns `false` if the injection should be blocked for the given hostname
- * otherwise returns `undefined` and performs necessary preparations for the injection
- * (like registering content scripts for scriptlet filters on Firefox)
- */
-async function injectCosmetics(details, config) {
-  const { bootstrap: isBootstrap = false, scriptletsOnly } = config;
-
-  try {
-    setup.pending && (await setup.pending);
-  } catch (e) {
-    console.error('[adblocker] not ready for cosmetic injection', e);
-    return;
-  }
-
-  const { tabId, frameId, parentFrameId, documentId, url } = details;
-
-  const parsed = parseWithCache(url);
-  const domain = parsed.domain || '';
-  const hostname = parsed.hostname || '';
-
-  if (__FIREFOX__ && scriptletsOnly && contentScripts.isRegistered(hostname)) {
-    return;
-  }
-
-  const options = store.get(Options);
-  // Checking the request url hostname
-  if (getPausedDetails(options, hostname)) {
-    return false;
-  }
-
-  // Checking the tab hostname to cover local iframes without a proper URL (about:blank, data:, etc.)
-  const tabHostname = tabStats.get(tabId)?.hostname;
-  if (tabHostname && getPausedDetails(options, tabHostname)) {
-    return false;
-  }
-
-  const engine = engines.get(engines.MAIN_ENGINE);
-
-  let ancestors = undefined;
-  if (SUBFRAME_SCRIPTING.enabled && typeof parentFrameId === 'number') {
-    if (__FIREFOX__) {
-      // On Firefox with content scripts API, we need to collect
-      // every scriptlets will potentially run on the hostname.
-      // Putting same values to `ancestors` enables adblocker to
-      // find all possible cases. The subframe constraint is
-      // validated by the `window.parent` property upon a script
-      // is executed.
-      ancestors = [{ domain, hostname }];
-    } else {
-      ancestors = framesHierarchy.ancestors(
-        { tabId, frameId, parentFrameId, documentId },
-        { domain, hostname },
-      );
-    }
-  }
-
-  // Domain specific cosmetic filters (scriptlets and styles)
-  // Execution: bootstrap, DOM mutations
-  {
-    const { matches } = engine.matchCosmeticFilters({
-      domain,
-      hostname,
-      url,
-      ancestors,
-
-      classes: config.classes,
-      hrefs: config.hrefs,
-      ids: config.ids,
-
-      // This needs to be done only once per frame
-      getInjectionRules: isBootstrap,
-      getExtendedRules: isBootstrap,
-      getRulesFromHostname: isBootstrap,
-
-      getPureHasRules: true,
-
-      // This will be done every time we get information about DOM mutation
-      getRulesFromDOM: !isBootstrap,
-      callerContext: { tabId },
-    });
-
-    const styleFilters = [];
-    const scriptFilters = [];
-
-    for (const { filter, exception } of matches) {
-      if (exception === undefined) {
-        if (filter.isScriptInject()) {
-          scriptFilters.push(filter);
-        } else {
-          styleFilters.push(filter);
-        }
-      }
-    }
-
-    if (isBootstrap) {
-      injectScriptlets(scriptFilters, hostname, details);
-    }
-
-    if (scriptletsOnly) {
-      return;
-    }
-
-    const { styles, extended } = engine.injectCosmeticFilters(styleFilters, {
-      url,
-      injectScriptlets: isBootstrap,
-      injectExtended: isBootstrap,
-      injectPureHasSafely: true,
-      allowGenericHides: false,
-      getBaseRules: false,
-    });
-
-    if (styles) {
-      injectStyles(styles, details);
-    }
-
-    if (extended && extended.length > 0) {
-      chrome.tabs.sendMessage(
-        tabId,
-        { action: 'evaluateExtendedSelectors', extended },
-        { frameId },
-      );
-    }
-  }
-
-  // Global cosmetic filters (styles only)
-  // Execution: bootstrap
-  if (isBootstrap) {
-    const { styles } = engine.getCosmeticsFilters({
-      domain,
-      hostname,
-      url,
-      getBaseRules: true,
-      getInjectionRules: false,
-      getExtendedRules: false,
-      getRulesFromDOM: false,
-      getRulesFromHostname: false,
-    });
-
-    injectStyles(styles, details);
-  }
-}
-
-/*
- * Cosmetic filtering
- */
-
-// Inject cosmetics on navigation committed (All platforms supported)
-chrome.webNavigation.onCommitted.addListener(
-  (details) => injectCosmetics(details, { bootstrap: true }),
-  { url: [{ urlPrefix: 'http://' }, { urlPrefix: 'https://' }] },
-);
-
-// Listen for requests from content scripts to inject
-// dynamic cosmetics (All platforms supported)
-chrome.runtime.onMessage.addListener((msg, sender, sendResponse) => {
-  if (msg.action === 'injectCosmetics' && sender.tab) {
-    // Generate details object for the sender argument
-
-    // Resolve url for the frame/subframe
-    // Local frames (about:blank, data:, etc.) do not have a proper URL
-    const url = !sender.url.startsWith('http') ? sender.origin : sender.url;
-
-    const details = {
-      url,
-      tabId: sender.tab.id,
-      frameId: sender.frameId,
-      documentId: sender.documentId,
-    };
-
-    injectCosmetics(details, msg).then(sendResponse);
-
-    return true;
-  }
-});
-
-if (__FIREFOX__) {
-  OptionsObserver.addListener('paused', function firefoxContentScriptScriptlets(paused) {
-    for (const hostname of Object.keys(paused)) {
-      contentScripts.unregister(hostname);
-    }
-  });
-
-  chrome.webNavigation.onBeforeNavigate.addListener(
-    (details) => {
-      injectCosmetics(details, { bootstrap: true, scriptletsOnly: true });
-    },
-    { url: [{ urlPrefix: 'http://' }, { urlPrefix: 'https://' }] },
-  );
-} else {
-  chrome.webRequest?.onResponseStarted.addListener(
-    (details) => {
-      if (details.tabId === -1) return;
-      if (details.type !== 'main_frame' && details.type !== 'sub_frame') return;
-
-      if (!details.documentId) return;
-
-      injectCosmetics(details, { bootstrap: true });
-    },
-    { urls: ['http://*/*', 'https://*/*'] },
-  );
-}
-
-/*
- * Network requests blocking - Firefox only
- */
-
-if (__FIREFOX__) {
-  function isTrusted(request, type) {
-    const options = store.get(Options);
-
-    // The request is from a tab that is paused
-    if (getPausedDetails(options, request.sourceHostname)) {
-      return true;
-    }
-
-    if (type === 'main_frame') {
-      return false;
-    }
-
-    return exceptions.getStatus(
-      options,
-      // Get exception for known tracker (metadata id) or by the request hostname (unidentified tracker)
-      trackerdb.getMetadata(request)?.id || request.hostname,
-      request.sourceHostname,
-    ).trusted;
-  }
-
-  function isMatchableRequest(details, request) {
-    // Extension context request
-    if (
-      (details.tabId === -1 && details.url.startsWith('moz-extension://')) ||
-      details.originUrl?.startsWith('moz-extension://')
-    ) {
-      return false;
-    }
-
-    // Engine not ready
-    if (setup.pending) {
-      console.error('[adblocker] not ready for network requests blocking');
-      return false;
-    }
-
-    // sourceHostname empty - for example for service workers
-    // Trusted request - for example from a paused tab
-    if (!request.sourceHostname || isTrusted(request, details.type)) {
-      return false;
-    }
-
-    return true;
-  }
-
-  chrome.webRequest.onBeforeRequest.addListener(
-    (details) => {
-      const request = Request.fromRequestDetails(details);
-      let result = undefined;
-
-      if (isMatchableRequest(details, request)) {
-        const engine = engines.get(engines.MAIN_ENGINE);
-        const { redirect, match } = engine.match(request);
-
-        if (match === true && details.type === 'main_frame') {
-          const options = store.get(Options);
-          const redirectUrl = getRedirectProtectionUrl(details.url, request.hostname, options);
-
-          return { redirectUrl };
-        } else if (redirect !== undefined) {
-          request.blocked = true;
-          // There's a possibility that redirecting to file URL can expose
-          // extension existence.
-          if (details.type !== 'xmlhttprequest') {
-            result = {
-              redirectUrl: chrome.runtime.getURL('rule_resources/redirects/' + redirect.filename),
-            };
-          } else {
-            result = { redirectUrl: redirect.dataUrl };
-          }
-        } else if (match === true) {
-          request.blocked = true;
-          result = { cancel: true };
-        }
-      }
-
-      updateTabStats(details.tabId, [request]);
-
-      return result;
-    },
-    { urls: ['<all_urls>'] },
-    ['blocking'],
-  );
-
-  chrome.webRequest.onHeadersReceived.addListener(
-    (details) => {
-      const request = Request.fromRequestDetails(details);
-      if (!isMatchableRequest(details, request)) return;
-
-      const engine = engines.get(engines.MAIN_ENGINE);
-
-      const htmlFilters = engine.getHtmlFilters(request);
-      if (htmlFilters.length !== 0) {
-        request.modified = true;
-        updateTabStats(details.tabId, [request]);
-        filterRequestHTML(chrome.webRequest.filterResponseData, request, htmlFilters);
-      }
-
-      if (details.type !== 'main_frame') return;
-
-      const cspPolicies = engine.getCSPDirectives(request);
-      if (!cspPolicies || cspPolicies.length === 0) return;
-
-      return updateResponseHeadersWithCSP(details, cspPolicies);
-    },
-    { urls: ['http://*/*', 'https://*/*'] },
-    ['blocking', 'responseHeaders'],
-  );
-}
+import './cosmetics.js';
+import './network.js';
