+++ src/background/dnr.js
@@ -19,7 +19,7 @@ import * as OptionsObserver from '/utils/options-observer.js';
 import { ENGINE_CONFIGS_ROOT_URL } from '/utils/urls.js';
 import { isFilterConditionAccepted } from '/utils/engines.js';
 
-import { UPDATE_ENGINES_DELAY } from './adblocker/index.js';
+import { UPDATE_ENGINES_DELAY } from './adblocker/engines.js';
 import { updateRedirectProtectionRules } from './redirect-protection.js';
 import { captureException } from '/utils/errors.js';
