+++ src/background/index.js
@@ -12,8 +12,8 @@
 import './onboarding.js';
 import './config.js';
 
-import './autoconsent.js';
 import './adblocker/index.js';
+import './autoconsent.js';
 import './cookies.js';
 import './custom-filters.js';
 import './element-picker.js';
